Raise error when JSON is invalid
Fix

+++ spec/jekyll-shields_io_spec.rb
@@ -118,7 +118,20 @@ RSpec.describe "Integration test" do
     end
   end
 
+  context "When invalid JSON is passed" do
+    it "fails with exception" do
+      t = Liquid::Template.new
+      expect { t.parse(
+        <<~EOT
+          {% shield_io {this is invalid!} %}
+      EOT
+      ) }.to raise_error Jekyll::ShieldsIO::ShieldConfigMalformedError
+    end
+  end
+
   after do
-    FileUtils.rm_r @cache_dir
+    if File.exists? @cache_dir
+      FileUtils.rm_r @cache_dir
+    end
   end
 end
